Remove extra files and rename
