--- client/src/components/application/dates/index.ts
@@ -1,5 +1,0 @@
-export * from "./applicationDates";
-export * from "./conceptDates";
-export * from "./stateApplicationDates";
-export * from "./completenessDates";
-export * from "./ApplicationDatesFlowchart";
