+++ client/src/router/DemosRouter.tsx
@@ -15,7 +15,7 @@ import { IconLibrary } from "pages/debug/IconLibrary";
 import { EventSandbox } from "pages/debug/EventSandbox";
 import { AuthDebugComponent } from "components/auth/AuthDebugComponent";
 import { isLocalDevelopment } from "config/env";
-import { ApplicationDatesFlowchart } from "components/application/dates/ApplicationDatesFlowchart";
+import { PhaseDatesSimulation } from "components/application/dates/phaseDatesSimulation";
 
 // 1) Shell that provides the layout and outlet for child routes
 function ProtectedShell() {
@@ -55,7 +55,7 @@ export const DemosRouter: React.FC = () => {
                     <Route path="icons" element={<IconLibrary />} />
                     <Route path="events" element={<EventSandbox />} />
                     <Route path="auth" element={<AuthDebugComponent />} />
-                    <Route path="dates" element={<ApplicationDatesFlowchart />} />
+                    <Route path="dates" element={<PhaseDatesSimulation />} />
                     <Route path="*" element={<div>404: Page Not Found</div>} />
                   </>
                 )}
